Support alternate MMS API URLs
MMS exists both as a hosted service as well as a suite of applications
that users can install within their own infrastructure. Users with large
clusters of MongoDB instances to monitor and backup will generally
prefer On-Premises MMS due to the savings in time/resources for
transferring backups as well as the costs savings from not paying
MongoDB for backup storage.

This small patch adds support for setting a custom MMS API URL to allow
the tool to be used against mms services other than the official
mongodb.com service.